# docs/openapi.yaml
openapi: 3.1.0
info:
  description: "This is API overview"
  version: 0.1.0
  title: PyrHouse API
  contact:
    email: warrmag7@gmail.com
  license:
    name: MIT
security:
  - bearerAuth: [] # use the same name as above    
# tags:
#   - name: users
#     description: Operations about user
#     externalDocs:
#       description: Find out more about...
#       url: http://swagger.io
paths:
...
  /items/categories:
    post:
      tags:
        - item category
      summary: Register user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  example: laptop       
                label:
                  type: string
                  example: Laptop
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemCategory"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - item category
      summary: Returns collection of users
      description: ...
      responses:
        "200":
          description: List all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemCategory"
...
  /users:
    post:
      tags:
        - users
      summary: Register user
      description: This can only be done by the logged in user with admin role.
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - role
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  example: password123
                fullname:
                  type: string
                  example: John Doe
                role:
                  $ref: "#/components/schemas/Role"
        description: Created user object
        required: true
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - users
      summary: Returns collection of users
      description: Returns a list of all users. Requires moderator role.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - users
      summary: Get user by ID
      description: Returns a single user. Requires user role.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags:
        - users
      summary: Update user
      description: Updates an existing user. Requires admin role.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: John Doe Updated
                password:
                  type: string
                  example: newpassword123
                role:
                  $ref: "#/components/schemas/Role"
        description: Updated user object
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request payload or user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}/points:
    parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
    patch:
      tags:
        - users
      summary: Update user points
      description: Updates points for an existing user. Requires admin role.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - points
              properties:
                points:
                  type: integer
                  description: New points value for the user
                  example: 100
        description: Points update request
        required: true
      responses:
        "200":
          description: Points updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User points updated successfully
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - users
      summary: Add or subtract user points
      description: Adds or subtracts points from an existing user. Use positive values to add points, negative to subtract. Requires admin role.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - points
              properties:
                points:
                  type: integer
                  description: Points to add (positive) or subtract (negative)
                  example: 10
        description: Points adjustment request
        required: true
      responses:
        "200":
          description: Points updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User points updated successfully
                  points:
                    type: integer
                    description: Current points after update
                    example: 110
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: pyrhouse-backend-f26ml.ondigitalocean.app/api
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for 
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Could not delete location
        details:
          type: string
...
    ItemCategory:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          example: asset 
        name:
          type: string
          example: laptop       
        label:
          type: string
          example: Laptop
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: johndoe
        fullname:
          type: string
          example: John Doe
        role:
          $ref: "#/components/schemas/Role"
        points:
          type: integer
          description: User's points
          example: 0
        active:
          type: boolean
          description: Czy konto jest aktywne
          example: true
        discord_id:
          type: string
          nullable: true
          description: ID użytkownika Discord (jeśli połączony)
          example: "123456789012345678"
        discord_username:
          type: string
          nullable: true
          description: Nazwa użytkownika Discord
          example: "user#1234"
        avatar_url:
          type: string
          nullable: true
          description: URL avatara Discord
          example: "https://cdn.discordapp.com/avatars/123/abc.png"
        auth_provider:
          type: string
          description: Metoda autoryzacji
          enum:
            - local
            - discord
          example: local
...
    Role:
      type: string
      description: Poziom uprawnień użytkownika
      enum:
        - user
        - moderator
        - admin
      example: user